[REFACTOR] Refactor handle post and upload record

+++ include/utils/utils.hpp
@@ -18,7 +18,7 @@ bool safeClose(int& fd);
 // bool setup_nonblocking(int fd);
 std::string trim(const std::string& str);
 int stoi(const std::string& str);
-unsigned int convertToSize(const std::string& str);
+size_t convertToSize(const std::string& str);
 std::vector<std::string> split(const std::string& str, char delim);
 bool isDirectory(const std::string& path);
 bool isFile(const std::string& path);
